feat: :sparkles: enhance spawnOn to place actors based upon load
