+++ actor/actor_system_test.go
@@ -33,7 +33,6 @@ import (
 	"reflect"
 	"strconv"
 	"strings"
-	"sync"
 	"testing"
 	"time"
 
@@ -123,93 +122,6 @@ func TestActorSystem(t *testing.T) {
 		assert.Error(t, err)
 		assert.Nil(t, sys)
 		require.ErrorIs(t, err, gerrors.ErrInvalidActorSystemName)
-	})
-	t.Run("With Spawn an actor when not System started", func(t *testing.T) {
-		ctx := context.TODO()
-		sys, _ := NewActorSystem("testSys", WithLogger(log.DiscardLogger))
-		actor := NewMockActor()
-		actorRef, err := sys.Spawn(ctx, "Test", actor)
-		assert.Error(t, err)
-		assert.ErrorIs(t, err, gerrors.ErrActorSystemNotStarted)
-		assert.Nil(t, sys.Metric(ctx))
-		assert.Nil(t, actorRef)
-		assert.Zero(t, sys.Uptime())
-	})
-	t.Run("With Spawn an actor when started", func(t *testing.T) {
-		ctx := context.TODO()
-		sys, _ := NewActorSystem("testSys", WithLogger(log.DiscardLogger))
-
-		// start the actor system
-		err := sys.Start(ctx)
-		assert.NoError(t, err)
-
-		pause.For(time.Second)
-
-		actor := NewMockActor()
-		actorRef, err := sys.Spawn(ctx, "Test", actor)
-		assert.NoError(t, err)
-		assert.NotNil(t, actorRef)
-
-		assert.NotZero(t, sys.Uptime())
-
-		// stop the actor after some time
-		pause.For(time.Second)
-		err = sys.Stop(ctx)
-		assert.NoError(t, err)
-	})
-	t.Run("With Spawn an actor with invalid actor name", func(t *testing.T) {
-		ctx := context.TODO()
-		sys, _ := NewActorSystem("testSys", WithLogger(log.DiscardLogger))
-
-		// start the actor system
-		err := sys.Start(ctx)
-		require.NoError(t, err)
-
-		actor := NewMockActor()
-		actorRef, err := sys.Spawn(ctx, "$omeN@me", actor)
-		require.Error(t, err)
-		assert.EqualError(t, err, "must contain only word characters (i.e. [a-zA-Z0-9] plus non-leading '-' or '_')")
-		assert.Nil(t, actorRef)
-
-		// stop the actor after some time
-		pause.For(time.Second)
-
-		t.Cleanup(
-			func() {
-				err = sys.Stop(ctx)
-				assert.NoError(t, err)
-			},
-		)
-	})
-	t.Run("With Spawn an actor already exist", func(t *testing.T) {
-		ctx := context.TODO()
-		sys, _ := NewActorSystem("test", WithLogger(log.DiscardLogger))
-
-		// start the actor system
-		err := sys.Start(ctx)
-		assert.NoError(t, err)
-
-		actor := NewMockActor()
-		ref1, err := sys.Spawn(ctx, "Test", actor)
-		assert.NoError(t, err)
-		assert.NotNil(t, ref1)
-
-		ref2, err := sys.Spawn(ctx, "Test", actor)
-		assert.NotNil(t, ref2)
-		assert.NoError(t, err)
-
-		// point to the same memory address
-		assert.True(t, ref1 == ref2)
-
-		// stop the actor after some time
-		pause.For(time.Second)
-
-		t.Cleanup(
-			func() {
-				err = sys.Stop(ctx)
-				assert.NoError(t, err)
-			},
-		)
 	})
 	t.Run("With RemoteActor/ActorOf with clustering enabled", func(t *testing.T) {
 		ctx := context.TODO()
@@ -1259,231 +1171,7 @@ func TestActorSystem(t *testing.T) {
 
 		require.NoError(t, sys.Stop(ctx))
 	})
-	t.Run("With SpawnNamedFromFunc when actor already exist", func(t *testing.T) {
-		ctx := context.TODO()
-		ports := dynaport.Get(1)
 
-		logger := log.DiscardLogger
-		host := "127.0.0.1"
-
-		newActorSystem, err := NewActorSystem(
-			"test",
-			WithLogger(logger),
-			WithRemote(remote.NewConfig(host, ports[0])))
-
-		require.NoError(t, err)
-
-		// start the actor system
-		err = newActorSystem.Start(ctx)
-		require.NoError(t, err)
-
-		receiveFn := func(_ context.Context, message proto.Message) error {
-			expected := &testpb.Reply{Content: "test spawn from func"}
-			assert.True(t, proto.Equal(expected, message))
-			return nil
-		}
-
-		actorName := "name"
-		actorRef, err := newActorSystem.SpawnNamedFromFunc(ctx, actorName, receiveFn)
-		assert.NoError(t, err)
-		assert.NotNil(t, actorRef)
-
-		// stop the actor after some time
-		pause.For(time.Second)
-
-		// send a message to the actor
-		require.NoError(t, Tell(ctx, actorRef, &testpb.Reply{Content: "test spawn from func"}))
-
-		newInstance, err := newActorSystem.SpawnNamedFromFunc(ctx, actorName, receiveFn)
-		require.NoError(t, err)
-		require.NotNil(t, newInstance)
-		require.True(t, newInstance.Equals(actorRef))
-
-		err = newActorSystem.Stop(ctx)
-		assert.NoError(t, err)
-	})
-	t.Run("With SpawnNamedFromFunc when actor name is invalid", func(t *testing.T) {
-		ctx := context.TODO()
-		ports := dynaport.Get(1)
-
-		logger := log.DiscardLogger
-		host := "127.0.0.1"
-
-		newActorSystem, err := NewActorSystem(
-			"test",
-			WithLogger(logger),
-			WithRemote(remote.NewConfig(host, ports[0])))
-
-		require.NoError(t, err)
-
-		// start the actor system
-		err = newActorSystem.Start(ctx)
-		require.NoError(t, err)
-
-		receiveFn := func(_ context.Context, message proto.Message) error {
-			expected := &testpb.Reply{Content: "test spawn from func"}
-			assert.True(t, proto.Equal(expected, message))
-			return nil
-		}
-
-		actorName := strings.Repeat("a", 256)
-		actorRef, err := newActorSystem.SpawnNamedFromFunc(ctx, actorName, receiveFn)
-		assert.Error(t, err)
-		assert.Nil(t, actorRef)
-
-		err = newActorSystem.Stop(ctx)
-		assert.NoError(t, err)
-	})
-	t.Run("With SpawnFromFunc (cluster/remote enabled)", func(t *testing.T) {
-		ctx := context.TODO()
-		nodePorts := dynaport.Get(3)
-		gossipPort := nodePorts[0]
-		clusterPort := nodePorts[1]
-		remotingPort := nodePorts[2]
-
-		logger := log.DiscardLogger
-		host := "127.0.0.1"
-
-		// define discovered addresses
-		addrs := []string{
-			net.JoinHostPort(host, strconv.Itoa(gossipPort)),
-		}
-
-		// mock the discovery provider
-		provider := new(mocks.Provider)
-		newActorSystem, err := NewActorSystem(
-			"test",
-			WithLogger(logger),
-			WithRemote(remote.NewConfig(host, remotingPort)),
-			WithCluster(
-				NewClusterConfig().
-					WithKinds(new(MockActor)).
-					WithPartitionCount(9).
-					WithReplicaCount(1).
-					WithPeersPort(clusterPort).
-					WithMinimumPeersQuorum(1).
-					WithDiscoveryPort(gossipPort).
-					WithDiscovery(provider)),
-		)
-		require.NoError(t, err)
-
-		provider.EXPECT().ID().Return("testDisco")
-		provider.EXPECT().Initialize().Return(nil)
-		provider.EXPECT().Register().Return(nil)
-		provider.EXPECT().Deregister().Return(nil)
-		provider.EXPECT().DiscoverPeers().Return(addrs, nil)
-		provider.EXPECT().Close().Return(nil)
-
-		// start the actor system
-		err = newActorSystem.Start(ctx)
-		require.NoError(t, err)
-
-		receiveFn := func(_ context.Context, message proto.Message) error {
-			expected := &testpb.Reply{Content: "test spawn from func"}
-			assert.True(t, proto.Equal(expected, message))
-			return nil
-		}
-
-		actorRef, err := newActorSystem.SpawnFromFunc(ctx, receiveFn)
-		assert.NoError(t, err)
-		assert.NotNil(t, actorRef)
-
-		// stop the actor after some time
-		pause.For(time.Second)
-
-		// send a message to the actor
-		require.NoError(t, Tell(ctx, actorRef, &testpb.Reply{Content: "test spawn from func"}))
-
-		t.Cleanup(
-			func() {
-				err = newActorSystem.Stop(ctx)
-				assert.NoError(t, err)
-				provider.AssertExpectations(t)
-			},
-		)
-	})
-	t.Run("With SpawnFromFunc with PreStart error", func(t *testing.T) {
-		ctx := context.TODO()
-		sys, _ := NewActorSystem("testSys", WithLogger(log.DiscardLogger))
-
-		// start the actor system
-		err := sys.Start(ctx)
-		assert.NoError(t, err)
-
-		receiveFn := func(_ context.Context, message proto.Message) error {
-			expected := &testpb.Reply{Content: "test spawn from func"}
-			assert.True(t, proto.Equal(expected, message))
-			return nil
-		}
-
-		preStart := func(ctx context.Context) error {
-			return errors.New("failed")
-		}
-
-		actorRef, err := sys.SpawnFromFunc(ctx, receiveFn, WithPreStart(preStart))
-		assert.Error(t, err)
-		assert.Nil(t, actorRef)
-
-		t.Cleanup(
-			func() {
-				err = sys.Stop(ctx)
-				assert.NoError(t, err)
-			},
-		)
-	})
-	t.Run("With SpawnFromFunc with PreStop error", func(t *testing.T) {
-		ctx := context.TODO()
-		sys, _ := NewActorSystem("testSys", WithLogger(log.DiscardLogger))
-
-		// start the actor system
-		err := sys.Start(ctx)
-		assert.NoError(t, err)
-
-		receiveFn := func(ctx context.Context, message proto.Message) error {
-			expected := &testpb.Reply{Content: "test spawn from func"}
-			assert.True(t, proto.Equal(expected, message))
-			return nil
-		}
-
-		postStop := func(ctx context.Context) error {
-			return errors.New("failed")
-		}
-
-		actorRef, err := sys.SpawnFromFunc(ctx, receiveFn, WithPostStop(postStop))
-		assert.NoError(t, err)
-		assert.NotNil(t, actorRef)
-
-		// stop the actor after some time
-		pause.For(time.Second)
-
-		// send a message to the actor
-		require.NoError(t, Tell(ctx, actorRef, &testpb.Reply{Content: "test spawn from func"}))
-
-		t.Cleanup(
-			func() {
-				assert.Error(t, sys.Stop(ctx))
-			},
-		)
-	})
-	t.Run("With SpawnFromFunc with actorSystem not started", func(t *testing.T) {
-		ctx := context.TODO()
-		sys, _ := NewActorSystem("testSys", WithLogger(log.DiscardLogger))
-
-		receiveFn := func(ctx context.Context, message proto.Message) error {
-			expected := &testpb.Reply{Content: "test spawn from func"}
-			assert.True(t, proto.Equal(expected, message))
-			return nil
-		}
-
-		preStart := func(ctx context.Context) error {
-			return errors.New("failed")
-		}
-
-		actorRef, err := sys.SpawnFromFunc(ctx, receiveFn, WithPreStart(preStart))
-		assert.Error(t, err)
-		assert.ErrorIs(t, err, gerrors.ErrActorSystemNotStarted)
-		assert.Nil(t, actorRef)
-	})
 	t.Run("With happy path Register", func(t *testing.T) {
 		ctx := context.TODO()
 		logger := log.DiscardLogger
@@ -1672,80 +1360,7 @@ func TestActorSystem(t *testing.T) {
 			},
 		)
 	})
-	t.Run("With Spawn with custom mailbox", func(t *testing.T) {
-		ctx := context.TODO()
-		actorSystem, _ := NewActorSystem("testSys", WithLogger(log.DiscardLogger))
 
-		// start the actor system
-		err := actorSystem.Start(ctx)
-		assert.NoError(t, err)
-
-		// wait for complete start
-		pause.For(time.Second)
-
-		// create the black hole actor
-		actor := NewMockActor()
-		pid, err := actorSystem.Spawn(ctx, "test", actor, WithMailbox(NewBoundedMailbox(10)))
-		assert.NoError(t, err)
-		assert.NotNil(t, pid)
-
-		// wait a while
-		pause.For(time.Second)
-		assert.EqualValues(t, 1, pid.ProcessedCount())
-		require.True(t, pid.IsRunning())
-
-		counter := 0
-		for i := 1; i <= 5; i++ {
-			require.NoError(t, Tell(ctx, pid, new(testpb.TestSend)))
-			counter = counter + 1
-		}
-
-		pause.For(time.Second)
-
-		assert.EqualValues(t, counter, pid.ProcessedCount()-1)
-		require.NoError(t, err)
-
-		err = actorSystem.Stop(ctx)
-		assert.NoError(t, err)
-	})
-	t.Run("With Spawn an actor already exist in cluster mode", func(t *testing.T) {
-		// create a context
-		ctx := context.TODO()
-		// start the NATS server
-		srv := startNatsServer(t)
-
-		// create and start system cluster
-		node1, sd1 := testNATs(t, srv.Addr().String())
-		peerAddress1 := node1.PeerAddress()
-		require.NotEmpty(t, peerAddress1)
-
-		// create and start system cluster
-		node2, sd2 := testNATs(t, srv.Addr().String())
-		peerAddress2 := node2.PeerAddress()
-		require.NotEmpty(t, peerAddress2)
-
-		pause.For(time.Second)
-
-		// create an actor on node1
-		actor := NewMockActor()
-		actorName := "actorID"
-		_, err := node1.Spawn(ctx, actorName, actor)
-		require.NoError(t, err)
-
-		pause.For(200 * time.Millisecond)
-
-		_, err = node2.Spawn(ctx, actorName, actor)
-		require.Error(t, err)
-		require.ErrorIs(t, err, gerrors.ErrActorAlreadyExists)
-
-		// free resource
-		require.NoError(t, node2.Stop(ctx))
-		assert.NoError(t, node1.Stop(ctx))
-		assert.NoError(t, sd2.Close())
-		assert.NoError(t, sd1.Close())
-		// shutdown the nats server gracefully
-		srv.Shutdown()
-	})
 	t.Run("With CoordinatedShutdown with ShouldFail strategy", func(t *testing.T) {
 		ctx := context.TODO()
 		executionCount := atomic.NewInt32(0)
@@ -2087,66 +1702,6 @@ func TestActorSystem(t *testing.T) {
 		require.NoError(t, sd3.Close())
 		// shutdown the nats server gracefully
 		srv.Shutdown()
-	})
-	t.Run("With Spawn with custom passivation", func(t *testing.T) {
-		ctx := context.TODO()
-		actorSystem, _ := NewActorSystem("testSys",
-			WithLogger(log.DiscardLogger))
-
-		require.NoError(t, actorSystem.Start(ctx))
-
-		pause.For(time.Second)
-
-		// create the actor path
-		pid, err := actorSystem.Spawn(ctx, "test", NewMockActor(),
-			WithPassivationStrategy(passivation.NewTimeBasedStrategy(passivateAfter)))
-		require.NoError(t, err)
-		assert.NotNil(t, pid)
-
-		// let us sleep for some time to make the actor idle
-		wg := sync.WaitGroup{}
-		wg.Add(1)
-		go func() {
-			pause.For(receivingDelay)
-			wg.Done()
-		}()
-		// block until timer is up
-		wg.Wait()
-		// let us send a message to the actor
-		err = Tell(ctx, pid, new(testpb.TestSend))
-		assert.Error(t, err)
-		assert.ErrorIs(t, err, gerrors.ErrDead)
-		assert.NoError(t, actorSystem.Stop(ctx))
-	})
-	t.Run("With Spawn with long lived", func(t *testing.T) {
-		ctx := context.TODO()
-		actorSystem, _ := NewActorSystem("testSys",
-			WithLogger(log.DiscardLogger))
-
-		require.NoError(t, actorSystem.Start(ctx))
-
-		pause.For(time.Second)
-
-		// create the actor path
-		pid, err := actorSystem.Spawn(ctx, "test", NewMockActor(),
-			WithLongLived())
-		require.NoError(t, err)
-		assert.NotNil(t, pid)
-
-		// let us sleep for some time to make the actor idle
-		wg := sync.WaitGroup{}
-		wg.Add(1)
-		go func() {
-			pause.For(receivingDelay)
-			wg.Done()
-		}()
-		// block until timer is up
-		wg.Wait()
-		// let us send a message to the actor
-		err = Tell(ctx, pid, new(testpb.TestSend))
-		assert.NoError(t, err)
-		assert.True(t, pid.IsRunning())
-		assert.NoError(t, actorSystem.Stop(ctx))
 	})
 	t.Run("With invalid remote config address", func(t *testing.T) {
 		remotingPort := dynaport.Get(1)[0]
@@ -2502,200 +2057,7 @@ func TestActorSystem(t *testing.T) {
 		err = sys.Stop(ctx)
 		require.NoError(t, err)
 	})
-	t.Run("SpawnOn happy path", func(t *testing.T) {
-		// create a context
-		ctx := context.TODO()
-		// start the NATS server
-		srv := startNatsServer(t)
 
-		// create and start system cluster
-		node1, sd1 := testNATs(t, srv.Addr().String())
-		peerAddress1 := node1.PeerAddress()
-		require.NotEmpty(t, peerAddress1)
-		require.NotNil(t, sd1)
-
-		// create and start system cluster
-		node2, sd2 := testNATs(t, srv.Addr().String())
-		peerAddress2 := node2.PeerAddress()
-		require.NotEmpty(t, peerAddress2)
-		require.NotNil(t, sd2)
-
-		// create and start system cluster
-		node3, sd3 := testNATs(t, srv.Addr().String())
-		peerAddress3 := node3.PeerAddress()
-		require.NotEmpty(t, peerAddress3)
-		require.NotNil(t, sd3)
-
-		pause.For(time.Second)
-
-		// create an actor on node1
-		actor := NewMockActor()
-		actorName := "actorID"
-		err := node1.SpawnOn(ctx, actorName, actor)
-		require.NoError(t, err)
-
-		pause.For(200 * time.Millisecond)
-
-		// either we can locate the actor or try to recreate it
-		_, err = node2.Spawn(ctx, actorName, actor)
-		require.Error(t, err)
-		require.ErrorIs(t, err, gerrors.ErrActorAlreadyExists)
-
-		// free resources
-		require.NoError(t, node2.Stop(ctx))
-		require.NoError(t, node1.Stop(ctx))
-		require.NoError(t, node3.Stop(ctx))
-
-		require.NoError(t, sd3.Close())
-		require.NoError(t, sd2.Close())
-		require.NoError(t, sd1.Close())
-
-		// shutdown the nats server gracefully
-		srv.Shutdown()
-	})
-	t.Run("SpawnOn when actor system not started", func(t *testing.T) {
-		// create a context
-		ctx := context.TODO()
-
-		actorSystem, _ := NewActorSystem("testSys",
-			WithLogger(log.DiscardLogger))
-
-		// create an actor on node1
-		actor := NewMockActor()
-		actorName := "actorID"
-		err := actorSystem.SpawnOn(ctx, actorName, actor)
-		require.Error(t, err)
-		require.ErrorIs(t, err, gerrors.ErrActorSystemNotStarted)
-	})
-	t.Run("SpawnOn when actor name is invalid", func(t *testing.T) {
-		ctx := context.TODO()
-		logger := log.DiscardLogger
-
-		// create the actor system
-		actorSystem, err := NewActorSystem(
-			"test",
-			WithLogger(logger),
-		)
-		// assert there are no error
-		require.NoError(t, err)
-
-		// start the actor system
-		err = actorSystem.Start(ctx)
-		assert.NoError(t, err)
-
-		pause.For(time.Second)
-
-		// create an actor on node1
-		actor := NewMockActor()
-		actorName := strings.Repeat("a", 256)
-		err = actorSystem.SpawnOn(ctx, actorName, actor)
-		require.Error(t, err)
-
-		err = actorSystem.Stop(ctx)
-		require.NoError(t, err)
-	})
-	t.Run("SpawnOn with random placement", func(t *testing.T) {
-		// create a context
-		ctx := context.TODO()
-		// start the NATS server
-		srv := startNatsServer(t)
-
-		// create and start system cluster
-		node1, sd1 := testNATs(t, srv.Addr().String())
-		peerAddress1 := node1.PeerAddress()
-		require.NotEmpty(t, peerAddress1)
-		require.NotNil(t, sd1)
-
-		// create and start system cluster
-		node2, sd2 := testNATs(t, srv.Addr().String())
-		peerAddress2 := node2.PeerAddress()
-		require.NotEmpty(t, peerAddress2)
-		require.NotNil(t, sd2)
-
-		// create and start system cluster
-		node3, sd3 := testNATs(t, srv.Addr().String())
-		peerAddress3 := node3.PeerAddress()
-		require.NotEmpty(t, peerAddress3)
-		require.NotNil(t, sd3)
-
-		pause.For(time.Second)
-
-		// create an actor on node1
-		actor := NewMockActor()
-		actorName := "actorID"
-		err := node1.SpawnOn(ctx, actorName, actor, WithPlacement(Random))
-		require.NoError(t, err)
-
-		pause.For(200 * time.Millisecond)
-
-		// either we can locate the actor or try to recreate it
-		_, err = node2.Spawn(ctx, actorName, actor)
-		require.Error(t, err)
-		require.ErrorIs(t, err, gerrors.ErrActorAlreadyExists)
-
-		// free resources
-		require.NoError(t, node2.Stop(ctx))
-		require.NoError(t, node1.Stop(ctx))
-		require.NoError(t, node3.Stop(ctx))
-
-		require.NoError(t, sd3.Close())
-		require.NoError(t, sd2.Close())
-		require.NoError(t, sd1.Close())
-
-		// shutdown the nats server gracefully
-		srv.Shutdown()
-	})
-	t.Run("SpawnOn with local placement", func(t *testing.T) {
-		// create a context
-		ctx := context.TODO()
-		// start the NATS server
-		srv := startNatsServer(t)
-
-		// create and start system cluster
-		node1, sd1 := testNATs(t, srv.Addr().String())
-		peerAddress1 := node1.PeerAddress()
-		require.NotEmpty(t, peerAddress1)
-		require.NotNil(t, sd1)
-
-		// create and start system cluster
-		node2, sd2 := testNATs(t, srv.Addr().String())
-		peerAddress2 := node2.PeerAddress()
-		require.NotEmpty(t, peerAddress2)
-		require.NotNil(t, sd2)
-
-		// create and start system cluster
-		node3, sd3 := testNATs(t, srv.Addr().String())
-		peerAddress3 := node3.PeerAddress()
-		require.NotEmpty(t, peerAddress3)
-		require.NotNil(t, sd3)
-
-		pause.For(time.Second)
-
-		// create an actor on node1
-		actor := NewMockActor()
-		actorName := "actorID"
-		err := node1.SpawnOn(ctx, actorName, actor, WithPlacement(Local))
-		require.NoError(t, err)
-
-		pause.For(200 * time.Millisecond)
-
-		// either we can locate the actor or try to recreate it
-		_, err = node2.Spawn(ctx, actorName, actor)
-		require.Error(t, err)
-		require.ErrorIs(t, err, gerrors.ErrActorAlreadyExists)
-
-		// free resources
-		require.NoError(t, node2.Stop(ctx))
-		require.NoError(t, node1.Stop(ctx))
-		require.NoError(t, node3.Stop(ctx))
-
-		require.NoError(t, sd3.Close())
-		require.NoError(t, sd2.Close())
-		require.NoError(t, sd1.Close())
-
-		// shutdown the nats server gracefully
-		srv.Shutdown()
-	})
 	t.Run("SpawnOn with single node cluster", func(t *testing.T) {
 		// create a context
 		ctx := context.TODO()
